Rewrite isLike
The primary problem was that the code didn't compile anymore, because
the typechecker wouldn't let us return a function of type, say,
(x: any) => x is boolean from isLike. Even with the workaround we've
been using up until now, we got this error:

    Type 'boolean' is not assignable to type 'T'.

The code did become shorter, less verbose and less error-prone, so
everything turned out better than expected, once I realized that eta
conversion does not hold in TypeScript with respect to types:

    return f;

does not work, but

    return (x: any): x is T => f(x);

does.