Make `is` total
`is(undefined)`, `is(null)` etc were runtime errors before this commit.

Now, we instead return a constant false type guard if the `type`
argument is a primitive. The intention is to resemble the semantics of
instanceof, but to return false instead of throwing. So, while

    undefined instanceof undefined

is a runtime error,

    is(undefined)(undefined)

is just false.