Add "Like" type guards for pseudo-primitive types
Members of Boolean, Number and String share a lot of the semantics with
members of boolean, number and string, respectively. For example,
members of String have all the methods that members of string do, such
as `match` and `charAt`.

However, they have completely different equality semantics:

               "foo"  ===            "foo" ; // true
    new String("foo") === new String("foo"); // false

There are other differences too, but this one on its own is enough to
keep the distinction between the primitive and the pseudo-primitive
types.

Resolves #10.